+++ src/OpenTelemetry.Exporter.OneCollector/CHANGELOG.md
@@ -14,6 +14,9 @@
   instrumentation key directly.
   ([#1037](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1037))
 
+* Added `RegisterPayloadTransmittedCallback` API on `OneCollectorExporter<T>`.
+  ([#1058](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1058))
+
 ## 0.1.0-alpha.1
 
 Released 2023-Feb-16